LayoutTests/media/track/track-cues-enter-exit.html

<!DOCTYPE html>
<html>
    <head>
        <meta http-equiv="Content-Type" content="text/html; charset=utf-8" />

        <script src=../media-file.js></script>
        <script src=../video-test.js></script>
    </head>
    <body>
        <p>Tests that TextTrack's cues are indexed and updated within 500ms during video playback. Test uses the enter and exits events on TextTrackCue.</p>
        <p>This test may fail occasionally since we cannot guarantee that cue events will fire within 250ms of their startTime or endTime.</p>
        <video controls>
            <track id="testTrack" src="captions-webvtt/cues-chrono-order.vtt" kind="captions">
        </video>
        <script>            
            
        var timeInterval = 0.5
        var cueCount = 0;
        var testTrack;

        function cueEntered()
        {
            consoleWrite("EVENT(enter)");

            var currentCue = testTrack.track.cues[cueCount];

            consoleWrite("This cue is the currently active cue:");
            testExpected(currentCue, testTrack.track.activeCues[0]);
            consoleWrite("Start time:");
            testExpected("video.currentTime", currentCue.startTime, '>=');
            testExpected("video.currentTime", currentCue.startTime + timeInterval, '<');
            consoleWrite("");
        }

        function cueExited()
        {
            consoleWrite("EVENT(exit)");

            var currentCue = testTrack.track.cues[cueCount];

            consoleWrite("End time:");
            testExpected("video.currentTime", currentCue.endTime, '>=');
            testExpected("video.currentTime", currentCue.endTime + timeInterval, '<');
            consoleWrite("");

            ++cueCount;
            if (cueCount == testTrack.track.cues.length)
                endTest();
        }

        waitForEvent('canplaythrough',
            function () 
            {
                testTrack = document.getElementById("testTrack");
                testExpected("testTrack.track.cues.length", 3);
                for (var i = 0; i < testTrack.track.cues.length; i++) {
                    testTrack.track.cues[i].addEventListener('enter', cueEntered);
                    testTrack.track.cues[i].addEventListener('exit', cueExited);
                }
                run("video.play()");
                consoleWrite("");
            }
        );

        findMediaElement();
        video.src = findMediaFile("video", "../content/test");
        
        </script>
    </body>
</html>
